feat: Add 'environment' option to babel-preset-udemy-website
'environment' can be 'legacy' or 'modern' and controls the set of browsers targeted by babel/preset-env
