+++ packages/babel-preset-udemy-website/tests/result/legacy/arrows.js
