+++ packages/babel-preset-udemy-website/tests/result/legacy/for-of-destructuring.js
