+++ packages/babel-preset-udemy-website/tests/result/legacy/object-spread.js
