+++ packages/babel-preset-udemy-website/tests/result/legacy/spread.js
