+++ packages/babel-preset-udemy-website/tests/result/modern/arrows.js
@@ -1,0 +1,11 @@
+"use strict";
+
+import "core-js/modules/es7.promise.finally";
+import "core-js/modules/es7.symbol.async-iterator";
+import "core-js/modules/web.timers";
+import "core-js/modules/web.immediate";
+import "core-js/modules/web.dom.iterable";
+import "regenerator-runtime/runtime";
+export default (function (it) {
+  return it;
+});
