+++ packages/babel-preset-udemy-website/tests/result/modern/template.js
@@ -1,0 +1,20 @@
+"use strict";
+
+import "core-js/modules/es7.promise.finally";
+import "core-js/modules/es7.symbol.async-iterator";
+import "core-js/modules/web.timers";
+import "core-js/modules/web.immediate";
+import "core-js/modules/web.dom.iterable";
+import "regenerator-runtime/runtime";
+
+function _templateObject() {
+  var data = babelHelpers.taggedTemplateLiteral(["1", "3"]);
+
+  _templateObject = function _templateObject() {
+    return data;
+  };
+
+  return data;
+}
+
+export default String.raw(_templateObject(), 2);
